feat: Separate unit tests from integration tests for local development
Story 0.3.2: Implement test separation to improve local developer experience

• Reorganize tests into unit/ and ci_only/ directories
• Unit tests run fast locally without Firebase dependencies
• CI-only tests skip locally, run in GitHub Actions with real Firebase configs
• Maintain all existing test functionality and coverage

Local development:
  flutter test test/unit    # Fast unit tests only (144 tests in ~8 seconds)

CI pipeline:
  flutter test             # All tests including Firebase integration
