+++ test/unit/core/widgets/environment_indicator_test.dart
