+++ test/unit/widget_test.dart
@@ -6,7 +6,7 @@ import 'package:play_with_me/features/auth/presentation/pages/login_page.dart';
 import 'package:play_with_me/features/auth/presentation/bloc/login/login_bloc.dart';
 import 'package:play_with_me/features/auth/presentation/widgets/auth_button.dart';
 import 'package:play_with_me/core/services/service_locator.dart';
-import 'helpers/test_helpers.dart';
+import '../helpers/test_helpers.dart';
 
 void main() {
   setUp(() async {
